2 cells

+++ run.sh
@@ -16,8 +16,8 @@ OUT_DIR=$RES_DIR"out/"
 IMG_DIR=$RES_DIR"images/"
 
 CONSTS_TABLE_FN="consts_table.txt" # input grid parameters
-MODEL_FN="microcin.pm"          # CTMC model 
-PROPERTIES_FN="microcin.csl"    # properties to test
+MODEL_FN="microcin-2cells.pm"          # CTMC model 
+PROPERTIES_FN="microcin-2cells.csl"    # properties to test
 TIME_FN="time.txt"              # time stamps for every prism run
 LOG_FN="prism.log"              # what prism said
 
@@ -36,13 +36,13 @@ LOG_FILE=$RES_DIR$LOG_FN
 DEFAULT_ARGS="-sor -fixdl"
 #CMD_LINE_ARGS=$@                             # additional command line arguments
 
-if [ -d $RES_DIR ]
-then
-  rm $RES_DIR -r
-fi
-mkdir $RES_DIR
-mkdir $CONSTS_DIR
-mkdir $OUT_DIR
+#if [ -d $RES_DIR ]
+#then
+#  rm $RES_DIR -r
+#fi
+#mkdir $RES_DIR
+#mkdir $CONSTS_DIR
+#mkdir $OUT_DIR
 
 # run the generative script
 $SCRIPT_DIR$GEN_SCRIPT $CONSTS_DIR $CONSTS_TABLE_FILE
@@ -74,12 +74,16 @@ for f in $CONSTS_DIR/*; do
   
   CONSTS="-const $(cat $f | tr -d '\n')"
   RES_FN="res.out.$num"
-  RESULTS="-exportresults $OUT_DIR$RES_FN"
+  RES_FILE=$OUT_DIR$RES_FN
+  RESULTS="-exportresults $RES_FILE"
   ARGS="$MODEL_FILE $PROPERTIES_FILE $CONSTS $RESULTS $DEFAULT_ARGS"
   
   #$SCRIPT_DIR/run.sh "$f"
   echo "prism "$ARGS
-  prism $ARGS | tee --append --ignore-interrupts $LOG_FILE
+  if [ ! -f $RES_FILE ]
+  then
+    prism $ARGS | tee --append --ignore-interrupts $LOG_FILE
+  fi
 
   #awk -v res_file=$RESULTS_FILE 'BEGIN{RS="\n\n"; ORS=""} { cnt++; print $0 > res_file"."cnt }' $RESULTS_FILE
